Fixed 2 bugs 1. ConvertToMono for 24-bit audio files was not working 2. If a recent file couldnt be found program was crashing

+++ WavEditCs/AudioProcessing.cs
@@ -273,8 +273,8 @@ namespace WavEditCs
                     }
                     else                    //stereo
                     {
-                        monoAudio.audioBuffer32[i] = (short)(aIn.audioBuffer32[i * 2] * 0.5f);
-                        monoAudio.audioBuffer32[i] += (short)(aIn.audioBuffer32[i * 2 + 1] * 0.5f);
+                        monoAudio.audioBuffer32[i] = (int)(aIn.audioBuffer32[i * 2] * 0.5f);
+                        monoAudio.audioBuffer32[i] += (int)(aIn.audioBuffer32[i * 2 + 1] * 0.5f);
                     }
                 }
